pyproject.toml
[project]
name = "tree-menu"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "likeinlife", email = "likeinlife@outlook.com" }
]
dependencies = [
    "django>=5.0.7",
    "pydantic-settings>=2.3.4",
    "django-split-settings>=1.3.2",
    "dishka>=1.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "ruff>=0.5.4",
    "mypy>=1.11.0",
    "pre-commit>=3.7.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Allow boolean type in func arguments
    "FBT001",
    # Allow PEP585 annotations like dict[str, str] 
    "FA102",
    # No blankline before class
    "D203",
    # Multiline string summary should start at the first line
    "D213",
    # Allow no-typehints
    "ANN",
    "UP008",
    # Disable doc-string rules
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D407",
    "D406",
    # Allow type: ignore
    "PGH003",
    # Useless
    "FBT",
    "FA100",
    "B904",
    "COM812",
    "TD003",
    "FIX002",
    "ANN401",
]
exclude = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "S", "ANN", "PLR"]
"**/migrations/*.py" = ["D", "S", "ANN", "PLR", "RUF"]

[tool.mypy]
ignore_missing_imports = true
follow_imports = "normal"
